backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
google.generativeai
gtts
shutils
